<Added requested changes and modified implementation flow>
