+++ app/controllers/application_controller.rb
@@ -50,26 +50,10 @@ class ApplicationController < ActionController::Base
     end
   end
 
-  def bbb_endpoint
-    Rails.configuration.bigbluebutton_endpoint
+  def bbb_connected?
+    Rails.configuration.bbb_connect
   end
-
-  def bbb_secret
-    Rails.configuration.bigbluebutton_secret
-  end
-
-  def remove_slash(s)
-    s.nil? ? nil : s.chomp("/")
-  end
-
-  def test_bbb_connection
-    unless Rails.configuration.loadbalanced_configuration
-      BigBlueButton::BigBlueButtonApi.new(remove_slash(bbb_endpoint), bbb_secret, "0.8").test_connection
-    end
-    true
-  rescue BigBlueButton::BigBlueButtonException
-    false
-  end
+  helper_method :bbb_connected?
 
   def meeting_name_limit
     MEETING_NAME_LIMIT
