+++ app/views/rooms/join.html.erb
@@ -23,7 +23,7 @@
           value: "#{current_user ? current_user.name : ''}",
           readonly: !current_user.nil? %>
 
-      <% if test_bbb_connection %>
+      <% if bbb_connected? %>
         <%= f.submit t("room.join"), class: "btn btn-primary btn-sm col-sm-3 form-control join-form" %>
       <% else %>
         <%= f.submit t("room.join"), disabled: true, class: "btn btn-primary btn-sm col-sm-3 form-control join-form" %>
