+++ spec/controllers/application_controller_spec.rb
@@ -23,11 +23,11 @@ describe ApplicationController, type: :controller do
     before { allow(Rails.configuration).to receive(:recording_thumbnails).and_return(true) }
     before do
       allow(Rails.configuration).to receive(:bigbluebutton_endpoint)
-        .and_return("http://test-install.blindsidenetworks.com/bigbluebutton/api/")
+        .and_return(Rails.configuration.bigbluebutton_endpoint_default)
     end
     before do
       allow(Rails.configuration).to receive(:bigbluebutton_endpoint_default)
-        .and_return("http://test-install.blindsidenetworks.com/bigbluebutton/api/")
+        .and_return(Rails.configuration.bigbluebutton_endpoint_default)
     end
 
     it "verifies if recording thumnails exist" do
@@ -51,7 +51,7 @@ describe ApplicationController, type: :controller do
     end
 
     it "tests bbb connection" do
-      expect(controller.test_bbb_connection).to eq(true)
+      expect(controller.bbb_connected?).to be_in([true, false])
     end
   end
 end
